確定日別による陽性者数の推移: Rename components/cards/PositiveNumberByDiagnosedDateCard.vue to components/index/CardsReference/PositiveNumberByDiagnosedDate/Card.vue
